+++ kayobe/cli/commands.py
@@ -229,6 +229,12 @@ class PhysicalNetworkConfigure(KayobeAnsibleMixin, VaultMixin, Command):
                            help="the Ansible group to apply configuration to")
         group.add_argument("--enable-discovery", action="store_true",
                            help="configure the network for hardware discovery")
+        group.add_argument("--interface-limit",
+                           help="limit the switch interfaces to be configured "
+                                "by interface name")
+        group.add_argument("--interface-description-limit",
+                           help="limit the switch interfaces to be configured "
+                                "by interface description")
         return parser
 
     def take_action(self, parsed_args):
@@ -236,6 +242,12 @@ class PhysicalNetworkConfigure(KayobeAnsibleMixin, VaultMixin, Command):
         extra_vars = {}
         if parsed_args.enable_discovery:
             extra_vars["physical_network_enable_discovery"] = True
+        if parsed_args.interface_limit:
+            extra_vars["physical_network_interface_limit"] = (
+                parsed_args.interface_limit)
+        if parsed_args.interface_description_limit:
+            extra_vars["physical_network_interface_description_limit"] = (
+                parsed_args.interface_description_limit)
         self.run_kayobe_playbook(parsed_args, "ansible/physical-network.yml",
                                  limit=parsed_args.group,
                                  extra_vars=extra_vars)